Fix deprecation warning (#937)
Put an `x` into the box(es) that apply:

- [X] This pull request fixes a warning.
- [ ] This pull request adds a feature.
- [ ] This pull request introduces breaking change.

When running example site with latest hugo version 0.135.5, a warning
appears:

```
WARN  deprecated: site config key paginate was deprecated in Hugo v0.128.0
and will be removed in a future release. Use pagination.pagerSize instead.
```

This PR solves that issues, both in code an in the documentation.
This PR also corrects and improves minor other topics in the
documentation.

Co-authored-by: Luiz F. A. de Prá <luizdepra@users.noreply.github.com>